# perf: add profiling annot in Scan_subcommand